add code v 25

+++ medical_management_BE/src/main/java/com/nhom2/qnu/payload/response/RevenueReportResponse.java
@@ -1,5 +1,7 @@
 package com.nhom2.qnu.payload.response;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class RevenueReportResponse {
     private String period; // YYYY-MM or YYYY-MM-DD
-    private Long totalRevenue;
+    private BigDecimal totalRevenue;
 }
