+++ medical_management_BE/src/main/java/com/nhom2/qnu/payload/response/ServiceUsageReportResponse.java
@@ -1,5 +1,7 @@
 package com.nhom2.qnu.payload.response;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,6 +11,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ServiceUsageReportResponse {
     private String serviceName;
-    private Long usageCount; // số lượt dùng
-    private Long totalRevenue; // tổng tiền thu được
+    private Long usageCount;
+    private BigDecimal totalRevenue;
 }
